[DRAFT 546681] Workaround NPE when closing editors asynchronously
Patch https://git.eclipse.org/r/c/125716 on the platform cleans up
some fields which asynchronous callbacks used to access for closing
the editors. At the time the callbacks are executed, getPage() and
getWorkbenchWindow() now return null.

This prevents the NPEs, but I'm not sure this is the correct fix as
Page.closeEditor() is no longer called (or has it already been called
when we end up in the async callback?).

Bug: 546681
Change-Id: I9a7831ea2643c7f7b8090633f524e2c035406fc8
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>